+++ neo/sound/snd_system.cpp
@@ -321,7 +321,7 @@ void idSoundSystemLocal::Init() {
 	}
 
 	// make a 16 byte aligned finalMixBuffer
-	finalMixBuffer = (float *) ( ( ( (int)realAccum ) + 15 ) & ~15 );
+	finalMixBuffer = (float *) ( ( ( (size_t)realAccum ) + 15 ) & ~15 );
 
 	graph = NULL;
